Moved over more stuff from the Settings class. On the control GUI buttons are no longer clickable if a schematic is not loaded, values default to 0/0/0. Exposed a few fields for some performance gains, no use encapsulating those anyway.

+++ src/main/java/com/github/lunatrius/schematica/proxy/ServerProxy.java
@@ -1,6 +1,6 @@
 package com.github.lunatrius.schematica.proxy;
 
-import com.github.lunatrius.core.util.vector.Vector3f;
+import com.github.lunatrius.core.util.vector.Vector3i;
 import com.github.lunatrius.schematica.world.SchematicWorld;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.server.MinecraftServer;
@@ -27,7 +27,7 @@ public class ServerProxy extends CommonProxy {
 	}
 
 	@Override
-	public boolean saveSchematic(EntityPlayer player, File directory, String filename, World world, Vector3f from, Vector3f to) {
+	public boolean saveSchematic(EntityPlayer player, File directory, String filename, World world, Vector3i from, Vector3i to) {
 		return false;
 	}
